Moves screener into lexer
