+++ src/zwreec/frontend/mod.rs
@@ -1,4 +1,3 @@
-pub mod screener;
 pub mod lexer;
 pub mod parser;
 pub mod ast;
